+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/TextFieldValidationBinderPage.java
@@ -16,29 +16,63 @@
 package com.vaadin.flow.component.textfield.tests.validation;
 
 import com.vaadin.flow.component.textfield.TextField;
+import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 
 @Route("vaadin-text-field/validation/binder")
-public class TextFieldValidationBinderPage
-        extends AbstractValidationBinderPage<TextField, String> {
+public class TextFieldValidationBinderPage extends AbstractValidationPage<TextField> {
     public static final String PATTERN_INPUT = "pattern-input";
     public static final String MIN_LENGTH_INPUT = "min-length-input";
     public static final String MAX_LENGTH_INPUT = "max-length-input";
+    public static final String EXPECTED_VALUE_INPUT = "expected-value-input";
+
+    public static final String REQUIRED_ERROR_MESSAGE = "The field is required";
+    public static final String UNEXPECTED_VALUE_ERROR_MESSAGE = "The field doesn't match the expected value";
+
+    public static class Bean {
+        private String property;
+
+        public String getProperty() {
+            return property;
+        }
+
+        public void setProperty(String property) {
+            this.property = property;
+        }
+    }
+
+    protected Binder<Bean> binder;
+
+    private String expectedValue;
 
     public TextFieldValidationBinderPage() {
+        super();
+
+        binder = new Binder<>(Bean.class);
+        binder.forField(field).asRequired(REQUIRED_ERROR_MESSAGE)
+                .withValidator(value -> value.equals(expectedValue),
+                        UNEXPECTED_VALUE_ERROR_MESSAGE)
+                .bind("property");
+
+        add(createInput(EXPECTED_VALUE_INPUT, "Set expected value", event -> {
+            expectedValue = event.getValue();
+        }));
+
         add(createInput(PATTERN_INPUT, "Set pattern", event -> {
             field.setPattern(event.getValue());
         }));
 
         add(createInput(MIN_LENGTH_INPUT, "Set min length", event -> {
-            int value = Integer.parseInt(event.getValue());
+            var value = Integer.parseInt(event.getValue());
             field.setMinLength(value);
         }));
 
         add(createInput(MAX_LENGTH_INPUT, "Set max length", event -> {
-            int value = Integer.parseInt(event.getValue());
+            var value = Integer.parseInt(event.getValue());
             field.setMaxLength(value);
         }));
+
+        addServerValidityStateControls();
     }
 
     protected TextField createField() {
